requirements.txt
streamlit==0.88.0
google-cloud-bigquery==2.30.1
sqlalchemy==1.4.23
langchain==1.3.1
openai==0.27.0
